cmp: Don't autocomplete on '<CR>'
Annoyingly, the default behavior of 'cmp' on NvChad is to select the
first auto-completion when Enter/Return is pressed. This is particularly
annoying at the end of a line. Rather than Return/Enter going to the
next line, it fills in the auto-completion and stays on the same line.

This change forces the user to press '<Tab>' to select an
auto-completion and then '<CR>' to commit the completion.

Signed-off-by: Chuck-R <github@chuck.cloud>